backend/requirements.txt
fastapi
uvicorn
pymysql
sqlalchemy
python-dotenv
passlib[bcrypt]
bcrypt==4.0.1
email-validator
cryptography
python-jose[cryptography]
python-multipart